ProductInclusion and ProductPrice Merge To Products Table

+++ PVMTrading_v1/Models/Product.cs
@@ -30,7 +30,7 @@ namespace PVMTrading_v1.Models
 
         [Required]
         [DataType(DataType.Currency)]
-        public decimal OriginalPrice { get; set; }
+        public double OriginalPrice { get; set; }
         //   public int LastChangeByEmployeeId { get; set; }
         public int Barcode { get; set; }
         //   public int WarrantyId { get; set; }
